♻️  Extract subscription classes for PSR 2
This allows us to rename the snake_case methods and to group them in a
more sensible manner.
